refactor: Move platform nodes to individual files
Implements ML-597
